update time implementation
