+++ shm/main.cpp
@@ -5,7 +5,7 @@
 //#include "observer.hpp" not ready yet
 #include "player.hpp"
 #include "ship.hpp"
-#include "time.hpp"
+#include "timeSHM.hpp"
 
 #include <iostream>
 #include <memory>
